+++ tests/packaging-utils/test_construct_contract_type_object.py
@@ -1,0 +1,7 @@
+from populus.packages.build import (
+    construct_contract_type_object,
+)
+
+
+def test_construct_contract_type_object():
+    assert False
